# openapi3.yaml
openapi: 3.0.1
info:
  title: replica-server
  description: A RESTful API for querying and saving geographic replicas (snapshot and deltas)
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
tags:
  - name: public
    description: the endpoints which are exposed to the client
  - name: private
    description: the endpoints which are exposed only to inner services
paths:
  /replica:
    get:
      operationId: getReplicas
      tags:
        - public
      summary: Get replicas by required and optional filters
      parameters:
        - in: query
          name: exclusive_from
          schema:
            $ref: '#/components/schemas/replicaDate'
          description: The lower limit of filtering by date. the filter is exclusive.
        - in: query
          name: to
          schema:
            $ref: '#/components/schemas/replicaDate'
          description: The upper limit of filtering by date. the filter is exclusive.
        - in: query
          name: sort
          description: The order in which the results are sorted based on their timestamp
          schema:
            type: string
            enum:
              - asc
              - desc
            default: desc
        - in: query
          name: geometry_type
          description: return results with the following geometry type - point, linestring or polygon
          required: true
          schema:
            $ref: '#/components/schemas/geometryType'
        - in: query
          name: replica_type
          description: return replicas of the following type - snapshot, delta
          required: true
          schema:
            $ref: '#/components/schemas/replicaType'
        - in: query
          name: layer_id
          description: return results from the following layer
          required: true
          schema:
            $ref: '#/components/schemas/layerId'
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: All the replicas returned by the search
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/replica'
        '400':
          $ref: '#/components/responses/BadRequest'
        5XX:
          $ref: '#/components/responses/UnexpectedError'
    post:
      operationId: postReplica
      tags:
        - private
      summary: Create a new replica
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/replica'
      security:
        - ApiKeyAuth: []
      responses:
        '201':
          description: Created
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        5XX:
          $ref: '#/components/responses/UnexpectedError'
    patch:
      operationId: patchReplica
      tags:
        - private
      summary: Update all replicas based on condition
      parameters:
        - $ref: '#/components/parameters/filterOptions'
      security:
        - ApiKeyAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/baseReplica'
      responses:
        '200':
          description: Updated
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        5XX:
          $ref: '#/components/responses/UnexpectedError'
    delete:
      operationId: deleteReplicas
      tags:
        - private
      summary: Delete replicas by filters
      parameters:
        - $ref: '#/components/parameters/filterOptions'
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/baseReplica'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        5XX:
          $ref: '#/components/responses/UnexpectedError'
  /replica/latest:
    get:
      operationId: getLatestReplica
      tags:
        - public
      summary: Get the latest replica for a layer and geometry type
      parameters:
        - in: query
          name: geometry_type
          description: return results with the following geometry type - point, linestring or polygon
          required: true
          schema:
            $ref: '#/components/schemas/geometryType'
        - in: query
          name: replica_type
          description: return replicas of the following type - snapshot, delta
          required: true
          schema:
            $ref: '#/components/schemas/replicaType'
        - in: query
          name: layer_id
          description: return results from the following layer
          required: true
          schema:
            $ref: '#/components/schemas/layerId'
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: All the replicas returned by the search
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/replica'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        5XX:
          $ref: '#/components/responses/UnexpectedError'
  /replica/{replicaId}:
    get:
      operationId: getReplicaById
      tags:
        - public
      summary: Get a replica by id
      parameters:
        - in: path
          name: replicaId
          schema:
            $ref: '#/components/schemas/replicaId'
          required: true
          description: id of the replica to get
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/replica'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        5XX:
          $ref: '#/components/responses/UnexpectedError'
  /replica/{replicaId}/file:
    post:
      operationId: postFile
      tags:
        - private
      summary: Create a new File for a Replica
      parameters:
        - in: path
          name: replicaId
          schema:
            $ref: '#/components/schemas/replicaId'
          required: true
          description: id of the replica to create file for
      security:
        - ApiKeyAuth: []
      responses:
        '201':
          description: Created
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        5XX:
          $ref: '#/components/responses/UnexpectedError'
  /layers:
    get:
      operationId: getLayers
      tags:
        - public
      summary: Get a list of available layers and their corresponding geometry types
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    layerId:
                      $ref: '#/components/schemas/layerId'
                    layerName:
                      $ref: '#/components/schemas/layerName'
                    geometryTypes:
                      type: array
                      items:
                        $ref: '#/components/schemas/geometryType'
        '400':
          $ref: '#/components/responses/BadRequest'
        5XX:
          $ref: '#/components/responses/UnexpectedError'
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
  parameters:
    filterOptions:
      name: filter
      in: query
      required: true
      schema:
        type: object
        properties:
          replica_type:
            $ref: '#/components/schemas/replicaType'
          geometry_type:
            $ref: '#/components/schemas/geometryType'
          layer_id:
            $ref: '#/components/schemas/layerId'
          sync_id:
            $ref: '#/components/schemas/syncId'
          is_hidden:
            type: boolean
          exclusive_from:
            $ref: '#/components/schemas/replicaDate'
          to:
            $ref: '#/components/schemas/replicaDate'
        additionalProperties: false
        minProperties: 1
    replicaType:
      in: query
      name: replica_type
      schema:
        $ref: '#/components/schemas/replicaType'
    layerId:
      in: query
      name: layer_id
      schema:
        $ref: '#/components/schemas/layerId'
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'
    Unauthorized:
      description: Unauthorized Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'
    UnexpectedError:
      description: Unexpected Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'
  schemas:
    geometryType:
      type: string
      enum:
        - point
        - linestring
        - polygon
    replicaType:
      type: string
      enum:
        - snapshot
        - delta
    replicaDate:
      type: string
      format: date-time
    replicaId:
      type: string
      format: uuid
    syncId:
      type: string
      format: uuid
    layerId:
      type: number
    layerName:
      type: string
    error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    baseReplica:
      type: object
      properties:
        timestamp:
          $ref: '#/components/schemas/replicaDate'
        layerId:
          type: number
          format: int32
        geometryType:
          $ref: '#/components/schemas/geometryType'
        isHidden:
          type: boolean
          writeOnly: true
          default: true
        bucketName:
          type: string
          minLength: 3
          maxLength: 63
          writeOnly: true
        replicaType:
          $ref: '#/components/schemas/replicaType'
        syncId:
          allOf:
            - writeOnly: true
            - $ref: '#/components/schemas/syncId'
        url:
          type: array
          items:
            type: string
            example: http://map.co/file.zip
          readOnly: true
    replica:
      allOf:
        - $ref: '#/components/schemas/baseReplica'
        - type: object
          properties:
            id:
              allOf:
                - writeOnly: true
                - $ref: '#/components/schemas/replicaId'